move loader to a separate file
